requirements.txt
# requirements.txt
# Work Order Matcher - Desktop Application Dependencies

# Google API packages (OAuth2 authentication pattern from reference project)
gspread>=5.7.0
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# Anthropic AI API
anthropic>=0.18.0

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Configuration management
python-dotenv>=0.19.0
pyyaml>=6.0

# HTTP requests
requests>=2.28.0

# GUI (tkinter included with Python standard library)
# tkinter - included with Python, no installation needed

# Optional: For future export functionality
# reportlab>=3.6.0  # Uncomment if PDF export needed
# Pillow>=9.0.0     # Uncomment if image processing needed 